Fix for bug#259
**Root cause**

The client constructor relies on the `getUser` method of the
`miscRequests` module, which requires location-specific URL to be passed
as one of its arguments.

Added a property `location` to the `clientOptions` object passed to the
client constructor. This `location` field will then be passed to the
`getUser` method within the constructor.